+++ backend/connector-integration/src/connectors.rs
@@ -103,3 +103,9 @@ pub use self::datatrans::Datatrans;
 
 pub mod authipay;
 pub use self::authipay::Authipay;
+
+pub mod silverflow;
+pub use self::silverflow::Silverflow;
+
+pub mod celero;
+pub use self::celero::Celero;